+++ ch05/instructions/loads/dload.go
@@ -6,7 +6,7 @@ import (
 )
 
 // Load Double variable from LocalVar table and push to opstack
-func _dload(frame *rtda.StackFrame, index uint8) {
+func _dload(frame *rtda.StackFrame, index uint) {
 	val := frame.LocalVars().GetDouble(index)
 	frame.OperandStack().PushDouble(val)
 }
